Fixed the issue with the SplitIndex
Fixed the issue with the SplitIndex. Now all objects are going to the
correct node. Added a AABB class and a Sphere class

+++ UntoldEngine4D/PhysicsEngine/CollisionEngine/BoundaryVolumeHierarchy/BVHTree/U4DBVHTree.h
@@ -95,9 +95,9 @@ namespace U4DEngine {
         
         U4DVector3n getLongestVolumeDimensionVector();
         
-        void setSplitIndex(float uSplitIndex);
+        void setSplitIndex(int uSplitIndex);
         
-        float getSplitIndex();
+        int getSplitIndex();
 
     };
